requirements.txt
django
gunicorn
django-heroku
requests
djangorestframework 
djongo
django-cors-headers
dnspython
pymongo[snappy,gssapi,srv,tls]